+++ src/DatasetBase/Support/AbstractCollection.php
@@ -67,6 +67,11 @@ abstract class AbstractCollection implements CollectionInterface
         $this->items = [];
     }
 
+    public function isEmpty(): bool
+    {
+        return $this->items === [];
+    }
+
     public function count(): int
     {
         return \count($this->items);